moved all printer manager functions
