+++ views/printerManagement.ejs
@@ -47,7 +47,7 @@
     <%- include(table, {
         table:{
             id:"printerTable", contentId:"printerList", headers: [
-                { title:"Sort Index" ,style:"min-width:50px;" },
+                { title:"Sort Index" ,style:"min-width:100px;" },
                 { title:"Name" ,style:"", wrap:" text-wrap" },
                 { title:"Actions" ,style:"min-width:250px;" },
                 { title:"Manage" ,style:"min-width:200px;" },
